修改年度汇总方法

+++ bill-gather/src/main/java/com/cloudfly/start/bill/mapper/BillBookAnnualSummaryMapper.java
@@ -5,6 +5,7 @@ import com.cloudfly.start.bill.entity.BillBookInfo;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
@@ -12,6 +13,6 @@ import java.util.Map;
 
 public interface BillBookAnnualSummaryMapper {
 
-    List<Map<String,Integer>> queryAnnualSummaryReportList(@Param("bookId") int bookId,
-                              @Param("startTime")Date startTime, @Param("endTime")Date endTime);
+    List<Map<String, BigDecimal>> queryAnnualSummaryReportList(@Param("bookId") int bookId,
+                                                               @Param("startTime")Date startTime, @Param("endTime")Date endTime);
 }
